refactor: 즐겨찾기 등록 시 해당 즐겨찾기 객체의 아이디를 반환하도록 변경

+++ src/main/java/nomad/backend/starred/Starred.java
@@ -12,7 +12,7 @@ import nomad.backend.member.Member;
 public class Starred {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer starredId;
+    private Long starredId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "memberId", nullable = false)
